# Remove unnecessary CDN in a Box waits