--- a/.github/actions/to-integration-tests/entrypoint.sh
+++ b/.github/actions/to-integration-tests/entrypoint.sh
@@ -72,7 +72,7 @@ start_traffic_vault() {
 		sed -i '0,/^update-ca-certificates/d' /etc/riak/prestart.d/00-config.sh;
 
 		# Do not try to source to-access.sh
-		sed -i '/to-access\.sh/d' /etc/riak/{prestart.d,poststart.d}/*
+		sed -i '/to-access\.sh\|^to-enroll/d' /etc/riak/{prestart.d,poststart.d}/*
 	BASH_LINES
 
 	DOCKER_BUILDKIT=1 docker build "$ciab_dir" -f "${ciab_dir}/traffic_vault/Dockerfile" -t "$trafficvault" 2>&1 |