--- a/infrastructure/cdn-in-a-box/docker-compose.traffic-ops-test.yml
+++ b/infrastructure/cdn-in-a-box/docker-compose.traffic-ops-test.yml
@@ -41,10 +41,10 @@ services:
 
   trafficops-perl:
     environment:
+      DIG_IP_RETRY: ${DIG_IP_RETRY:-0}
       LOAD_TRAFFIC_OPS_DATA: ${LOAD_TRAFFIC_OPS_DATA:-false}
 
 volumes:
   junit:
   shared:
     external: false
-