--- a/infrastructure/cdn-in-a-box/variables.env
+++ b/infrastructure/cdn-in-a-box/variables.env
@@ -122,3 +122,4 @@ AUTO_SNAPQUEUE_ENABLED=true
 AUTO_SNAPQUEUE_SERVERS=trafficops,trafficops-perl,trafficmonitor,trafficrouter,trafficvault,edge,mid-01,mid-02
 AUTO_SNAPQUEUE_POLL_INTERVAL=2
 AUTO_SNAPQUEUE_ACTION_WAIT=2
+LOAD_TRAFFIC_OPS_DATA=${LOAD_TRAFFIC_OPS_DATA:-true}