<!--
************ STOP!! ************
If this Pull Request is intended to fix a security vulnerability, DO NOT submit it! Instead, contact
the Apache Software Foundation Security Team at security@trafficcontrol.apache.org and follow the
guidelines at https://www.apache.org/security/ regarding vulnerability disclosure.
-->
## What does this PR (Pull Request) do?
<!-- Explain the changes you made here. If this fixes an Issue, identify it by
replacing the text in the checkbox item with the Issue number e.g.

- [x] This PR fixes #9001 OR is not related to any Issue

^ This will automatically close Issue number 9001 when the Pull Request is
merged (The '#' is important).

Be sure you check the box properly, see the "The following criteria are ALL
met by this PR" section for details.
-->

* [x] This PR is not related to any Issue <!-- You can check for an issue here: https://github.com/apache/trafficcontrol/issues -->
* This PR makes CDN in a Box start up faster.
    - Lets you skip things waiting for the enroller initial data load to complete. This saves about 24 seconds.
    - Lets you set `DIG_IP_RETRY` from `set-to-ips-from-dns.sh` to any value, letting you optionally skip this step . When set to 0, this saves 30 seconds per missing service. So, in the case of the Traffic Ops API tests, where the `trafficportal`, `trafficmonitor`, and `trafficrouter` are not included, this saves 90 seconds.
    - Traffic Vault now waits to enroll itself until after Riak finishes starting up. Saves about 24 seconds (the amount of time the enroller initial data load takes).
 
 Being able to skip the enroller initial data load is a bugfix for the Traffic Ops API tests. Right now, if you start the `integration` service (from `docker-compose.traffic-ops-test.yml`) when starting all of the other CDN in a Box services, the Traffic Ops API v1 tests fail (but v2 and v3 succeed) at the end of `TestCacheGroupsDeliveryServices` because the enroller continues seeding data after `Teardown()` truncates all the tables.
```
    types_test.go:96: ---- DeleteTestTypes ----
    types_test.go:108: cannot DELETE Type by name: 400 Bad Request[400] - Error requesting Traffic Ops https://trafficops.infra.ciab.test:443/api/1.3/types/135 {"alerts":[{"text":"cannot delete type because it is being used by a cachegroup","level":"error"}]}
         - {[]}
    types_test.go:117: expected Type name: ORG_LOC to be deleted
    types_test.go:108: cannot DELETE Type by name: 400 Bad Request[400] - Error requesting Traffic Ops https://trafficops.infra.ciab.test:443/api/1.3/types/136 {"alerts":[{"text":"cannot delete type because it is being used by a cachegroup","level":"error"}]}
         - {[]}
    types_test.go:117: expected Type name: MID_LOC to be deleted
    types_test.go:108: cannot DELETE Type by name: 400 Bad Request[400] - Error requesting Traffic Ops https://trafficops.infra.ciab.test:443/api/1.3/types/145 {"alerts":[{"text":"cannot delete type because it is being used by a cachegroup","level":"error"}]}
         - {[]}
    types_test.go:117: expected Type name: EDGE_LOC to be deleted
--- FAIL: TestCacheGroupsDeliveryServices (8.69s)
```

## Which Traffic Control components are affected by this PR?
<!-- Please delete all components from this list that are NOT affected by this
Pull Request. Also, feel free to add the name of a tool or script that is
affected but not on the list.

Additionally, if this Pull Request does NOT affect documentation, please
explain why documentation is not required. -->

- CDN in a Box
- Documentation
- CI tests

## What is the best way to verify this PR?
<!-- Please include here ALL the steps necessary to test your Pull Request. If
it includes tests (and most should), outline here the steps needed to run the
tests. If not, lay out the manual testing procedure and please explain why
tests are unnecessary for this Pull Request. -->
* Make sure the Traffic Ops API tests pass when `dig` retries and enroller initial data loading is disabled:
    ```
    export DIG_IP_RETRY=0 LOAD_TRAFFIC_OPS_DATA=false;
    docker-compose up --no-deps -d edge dns db smtp trafficops-perl trafficops trafficvault;
    docker-compose -f docker-compose.traffic-ops-test.yml -f docker-compose.yml up integration;
    ```

* Make sure CDN in a Box still works when started normally.

## If this is a bug fix, what versions of Traffic Control are affected?
<!-- If this PR fixes a bug, please list here all of the affected versions - to
the best of your knowledge. It's also pretty helpful to include a commit hash
of where 'master' is at the time this PR is opened (if it affects master),
because what 'master' means will change over time. For example, if this PR
fixes a bug that's present in master (at commit hash '1df853c8'), in v4.0.0,
and in the current 4.0.1 Release candidate (e.g. RC1), then this list would
look like:

- master (1df853c8)
- 4.0.0
- 4.0.1 (RC1)

If you don't know what other versions might have this bug, AND don't know how
to find the commit hash of 'master', then feel free to leave this section
blank (or, preferably, delete it entirely).
 -->
For the Traffic Ops API tests:
 
- master (f6a8056513)
- 5.0.0 (RC1)

## The following criteria are ALL met by this PR
<!-- Check the boxes to signify that the associated statement is true. To
"check a box", replace the space inside of the square brackets with an 'x'.
e.g.

- [ x] <- Wrong
- [x ] <- Wrong
- [] <- Wrong
- [*] <- Wrong
- [x] <- Correct!

-->

- [x] This PR fixes tests
- [x] This PR includes documentation
- [x] An update to CHANGELOG.md is not necessary
- [x] This PR includes any and all required license headers
- [x] This PR **DOES NOT FIX A SERIOUS SECURITY VULNERABILITY** (see [the Apache Software Foundation's security guidelines](https://www.apache.org/security/) for details)

<!--
Licensed to the Apache Software Foundation (ASF) under one
or more contributor license agreements.  See the NOTICE file
distributed with this work for additional information
regarding copyright ownership.  The ASF licenses this file
to you under the Apache License, Version 2.0 (the
"License"); you may not use this file except in compliance
with the License.  You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing,
software distributed under the License is distributed on an
"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
KIND, either express or implied.  See the License for the
specific language governing permissions and limitations
under the License.
-->